Enable strict warning flags and fix gen script issue
Add compiler warning flags as used in KBase.

Also fix the --build-test issue of gen.py (checkout KBase commit logs to see details of this fix).

+++ ezio/buffer_posix.cpp
@@ -25,7 +25,7 @@ ssize_t ReadFDInVec(int fd, Buffer& buf)
         return -1;
     }
 
-    if (size_read <= buf.writable_size()) {
+    if (static_cast<size_t>(size_read) <= buf.writable_size()) {
         buf.EndWrite(static_cast<size_t>(size_read));
     } else {
         auto writable = buf.writable_size();
